+++ src/config/src/ProviderConfig.php
@@ -35,23 +35,37 @@ class ProviderConfig
     public static function load(): array
     {
         if (! static::$privoderConfigs) {
-            $config = [];
-            $providers = Composer::getMergedExtra('hyperf')['config'];
-            foreach ($providers ?? [] as $provider) {
-                if (is_string($provider) && class_exists($provider) && method_exists($provider, '__invoke')) {
-                    $providerConfig = (new $provider())();
-                    $config = array_merge_recursive($config, $providerConfig);
-                }
-            }
-
-            static::$privoderConfigs = $config;
-            unset($config, $providerConfig);
+            $providers = Composer::getMergedExtra('hyperf')['config'] ?? [];
+            static::$privoderConfigs = static::loadProviders($providers);
         }
         return static::$privoderConfigs;
     }
 
-    public function clear(): void
+    public static function clear(): void
     {
         static::$privoderConfigs = [];
+    }
+
+    protected static function loadProviders(array $providers): array
+    {
+        $providerConfigs = [];
+        foreach ($providers ?? [] as $provider) {
+            if (is_string($provider) && class_exists($provider) && method_exists($provider, '__invoke')) {
+                $providerConfigs[] = (new $provider())();
+            }
+        }
+
+        return static::merge(...$providerConfigs);
+    }
+
+    protected static function merge(...$arrays): array
+    {
+        $result = array_merge_recursive(...$arrays);
+        if (isset($result['dependencies'])) {
+            $dependencies = array_column($arrays, 'dependencies');
+            $result['dependencies'] = array_merge(...$dependencies);
+        }
+
+        return $result;
     }
 }